+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/client/EcCapsuleClientImpl.java
@@ -7,10 +7,11 @@ import ee.cyber.cdoc2.client.model.Capsule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.nio.ByteBuffer;
 import java.security.GeneralSecurityException;
 import java.security.interfaces.ECPublicKey;
 import java.util.Optional;
+
+import static ee.cyber.cdoc2.crypto.EllipticCurve.*;
 
 
 @SuppressWarnings("java:S2139")
@@ -43,8 +44,9 @@ public class EcCapsuleClientImpl implements EcCapsuleClient {
         }
 
         var capsuleType = switch (curve) {
-            case SECP384R1 -> Capsule.CapsuleTypeEnum.ECC_SECP384R1;
             case SECP256R1 -> Capsule.CapsuleTypeEnum.ECC_SECP256R1;
+            case SECP384R1 -> Capsule.CapsuleTypeEnum.ECC_SECP384R1;
+            case SECP521R1 -> Capsule.CapsuleTypeEnum.ECC_SECP521R1;
             default -> throw new IllegalArgumentException("Unsupported EC curve " + curve);
         };
 
@@ -65,11 +67,14 @@ public class EcCapsuleClientImpl implements EcCapsuleClient {
                 Capsule capsule = capsuleOptional.get();
 
                 return switch (capsule.getCapsuleType()) {
-                    case ECC_SECP384R1 -> Optional.of(
-                        EllipticCurve.SECP384R1.decodeFromTls(ByteBuffer.wrap(capsule.getEphemeralKeyMaterial()))
-                    );
                     case ECC_SECP256R1 -> Optional.of(
-                        EllipticCurve.SECP256R1.decodeFromTls(ByteBuffer.wrap(capsule.getEphemeralKeyMaterial()))
+                        ECKeys.decodeEcPublicKeyFromTls(SECP256R1, capsule.getEphemeralKeyMaterial())
+                    );
+                    case ECC_SECP384R1 -> Optional.of(
+                        ECKeys.decodeEcPublicKeyFromTls(SECP384R1, capsule.getEphemeralKeyMaterial())
+                    );
+                    case ECC_SECP521R1 -> Optional.of(
+                        ECKeys.decodeEcPublicKeyFromTls(SECP521R1, capsule.getEphemeralKeyMaterial())
                     );
                     default -> throw new ExtApiException("Unsupported capsule type " + capsule.getCapsuleType());
                 };