+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/crypto/PemTools.java
@@ -19,7 +19,6 @@ import java.security.cert.CertificateEncodingException;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
-import java.security.interfaces.ECKey;
 import java.security.interfaces.ECPrivateKey;
 import java.security.interfaces.RSAPublicKey;
 import java.util.AbstractMap;
@@ -170,18 +169,10 @@ public final class PemTools {
 
         publicKey = keyPair.getPublic();
         if (KeyAlgorithm.isEcKeysAlgorithm(publicKey.getAlgorithm())) {
-            String oid = getCurveOid((ECKey) publicKey);
+            var curve = EllipticCurve.forPubKey(publicKey);
 
-            if (SECP_384_OID.equals(oid)) {
-                if (!ECKeys.isECSecp384r1(keyPair)) {
-                    throw new InvalidKeyException("Not an EC keypair with secp384r1 curve");
-                }
-            } else if (SECP_256_OID.equals(oid)) {
-                if (!ECKeys.isECSecp256r1(keyPair)) {
-                    throw new InvalidKeyException("Not an EC keypair with secp256r1 curve");
-                }
-            } else {
-                throw new InvalidKeyException("Unknown EC keypair OID: " + oid);
+            if (!isECKeyPairForCurve(curve, keyPair)) {
+                throw new InvalidKeyException("Unsupported key algorithm " + publicKey.getAlgorithm());
             }
         } else if (KeyAlgorithm.isRsaKeysAlgorithm(publicKey.getAlgorithm())) {
             // all RSA keys are considered good. Shorter will fail during encryption as OAEP takes some space